feat(prDetailView): add label filter and hide on blur
The list of labels in the label-manager is now filtered by the text
entered in the filter input box.

The label-manager drop-down now hides if you click away from it.

Fixed repeater for the checked label list to "track by" label.name,
since before it wasn't updating array correctly.  I suspect this is
due to the filter and orderBy filters reusing array elements.

Closes #6
Closes #13